requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
werkzeug==2.3.7
